Rename course outline store file
