+++ assets/blocks/course-outline/index.js
@@ -4,5 +4,5 @@
 export { default as OutlineBlock } from './outline-block';
 export { default as ModuleBlock } from './module-block';
 export { default as LessonBlock } from './lesson-block';
-import './store';
+import './course-outline-store';
 import './status-preview/status-store';
